Move deactivate-snapshot to a dedicated package with minimal imports
